requirements/base.txt
Django>=1.8.0,<1.9
GitPython>=1.0
Markdown>=2.6
PyYAML==3.10
boto==2.9.5
celery>=3.1
dj-database-url>=0.3
django-auth-ldap>=1.2
django-extensions>=1.5
django-filter>=0.9
django-guardian==1.2.6
django-model-utils>=2.0
djangorestframework>=3.0
envoy>=0.0.2
psutil>=2.1
requests>=2.4
salt==2014.1.7
